requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
base58==2.1.1
billiard==3.6.4.0
bitarray==2.6.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.12.0
Deprecated==1.2.13
Django==4.0.8
django-celery-beat==2.3.0
django-timezone-field==5.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.1
hexbytes==0.3.0
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.14.0
kombu==5.2.4
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.23.3
packaging==21.3
parsimonious==0.8.1
prompt-toolkit==3.0.31
protobuf==3.19.5
psycopg2==2.9.3
psycopg2-binary==2.9.4
pycoingecko==3.0.0
pycryptodome==3.15.0
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.2
toolz==0.12.0
twilio==7.14.2
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.11
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.31.1
websockets==9.1
whitenoise==6.2.0
wrapt==1.14.1
yarl==1.8.1
